feat: Add jest (#30)
Fixes #11 

In the process of writing a test for `replaceWords` I realized that it didn't actually work in the case of multiple entries -- so I changed that function, too. That's what tests are for I guess 😆 	

Also removed the cypress directory from tsconfig so that types work fine for jest tests.