fix(gatsby-theme-minimal-blog): Use correct h1 on all pages (#519)
This replaces the incorrect h1 in the header with a normal div and sets all actual page titles (e.g. blog post title, page title, titles on /tags overview and /tag page) to h1. To keep accessibility the homepage has a visually hidden h1 with the title (previously provided by the one in the header). This change also changes the font sizes of the heading styles a bit.

With this change the titles will appear larger beyond > 1024px. If you don't want that behavior, you can override the styles in the Theme UI config.

Co-authored-by: LekoArts <lekoarts@gmail.com>